feat: reemplazar el esquema de validación de clientes con soporte para traducciones
